WindowsControl/requirements.txt
# Core dependencies
requests>=2.32.0
pillow>=11.0.0
mss>=9.0.0
psutil>=6.0.0
pywin32>=311
pyautogui>=0.9.54
comtypes>=1.4.0

# MCP client
fastmcp>=2.14.2

# CLI
rich>=14.0.0
prompt-toolkit>=3.0.0

# Utilities
python-dotenv>=1.0.0

# Gesture / Speech WebSocket client
websockets>=13.0

# Text-to-Speech (Kokoro TTS)
kokoro-onnx>=0.5.0
soundfile
